troca de primitiva

+++ Ex/Ex007.py
@@ -1,8 +1,8 @@
 # Desenvolva um programa que leia as duas notas de um aluno, calcule e msotre a 
 # média
 
-n1 = int(input('Digite a primeira nota do aluno ciclano: '))
-n2 = int(input('Digite a segunda nota: '))
+n1 = float(input('Digite a primeira nota do aluno ciclano: '))
+n2 = float(input('Digite a segunda nota: '))
 print(
     f"""
     Cilcano obteve a primeira nota: {n1}
